[Mod] API 추가 및 정리 - API 파일 통합
