+++ src/pages/poke-dex/list/PokeDexListItem.tsx
@@ -8,11 +8,11 @@ import { PAGE_URL } from 'consts/common';
 
 interface IPokemonCard {
   pokemon: PokeDexList[number];
-  listItemRef?: React.RefObject<HTMLAnchorElement>;
+  fetchTriggerRef?: React.RefObject<HTMLAnchorElement>;
 }
-const PokeDexListItem = ({ pokemon, listItemRef }: IPokemonCard) => {
+const PokeDexListItem = ({ pokemon, fetchTriggerRef }: IPokemonCard) => {
   return (
-    <PokeDexListItemWrapper to={`${PAGE_URL.POKE_DEX}/${pokemon.id}`} ref={listItemRef}>
+    <PokeDexListItemWrapper to={`${PAGE_URL.POKE_DEX}/${pokemon.id}`} ref={fetchTriggerRef}>
       <PokeDexListItemImg src={pokemon.img} alt={pokemon.name} />
       <PokeDexListItemName>
         No.{pokemon.id}
